Address @proksch comments
