+++ plugins/ingest-artifact-completion/pom.xml
@@ -15,7 +15,6 @@
             <groupId>eu.fasten-project</groupId>
             <artifactId>pom-analyzer</artifactId>
             <version>0.0.8-SNAPSHOT</version>
-            <scope>compile</scope>
         </dependency>
     </dependencies>
 </project>
