--- plugins/ingest-artifact-completion/src/main/java/eu/f4sten/ingestartifactcompletion/IngestArtifactCompletionInjectorConfig.java
@@ -1,49 +1,0 @@
-/*
- * Copyright 2022 Delft University of Technology
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package eu.f4sten.ingestartifactcompletion;
-
-import com.google.inject.Binder;
-import com.google.inject.Provides;
-import eu.f4sten.infra.IInjectorConfig;
-import eu.f4sten.infra.InjectorConfig;
-import eu.f4sten.infra.json.JsonUtils;
-import eu.f4sten.infra.utils.PostgresConnector;
-import eu.f4sten.infra.utils.Version;
-import eu.f4sten.pomanalyzer.utils.DatabaseUtils;
-import org.jooq.SQLDialect;
-import org.jooq.impl.DSL;
-
-@InjectorConfig
-public class IngestArtifactCompletionInjectorConfig implements IInjectorConfig {
-    private IngestArtifactCompletionArgs args;
-
-    public void IngestArtifactCompletionArgs(IngestArtifactCompletionArgs args) {
-        this.args = args;
-    }
-
-    @Override
-    public void configure(Binder binder) {
-        binder.bind(IngestArtifactCompletionArgs.class).toInstance(args);
-    }
-
-    @Provides
-    public DatabaseUtils bindDatabaseUtils(PostgresConnector pc, JsonUtils json, Version version) {
-        var c = pc.getNewConnection();
-        var dslContext = DSL.using(c, SQLDialect.POSTGRES);
-        return new DatabaseUtils(dslContext, json, version);
-    }
-}
